refactor: clean up the proxy bypass
