+++ api/src/utils/passthough-proxy.ts
@@ -1,16 +1,17 @@
 import http, { IncomingHttpHeaders } from "node:http";
+import { PrepareRequestFunctionOpts, PrepareRequestFunctionResult } from "proxy-chain";
 
 // Headers that are hop-by-hop and should not be forwarded RFC 2616, Section 13.5.1
-const hopByHopHeaders = [
-  'connection',
-  'keep-alive',
-  'proxy-authenticate',
-  'proxy-authorization',
-  'te',
-  'trailers',
-  'transfer-encoding',
-  'upgrade',
-];
+export const hopByHopHeaders = new Set([
+  "connection",
+  "proxy-authenticate",
+  "proxy-authorization",
+  "keep-alive",
+  "te",
+  "trailers",
+  "transfer-encoding",
+  "upgrade",
+]);
 
 /**
  * Creates a simple http proxy which reverse proxies the original request to the original host.
@@ -27,7 +28,7 @@ export const PassthroughServer = http.createServer((clientReq, clientRes) => {
   const proxyHeaders: IncomingHttpHeaders = {};
   for (const [key, value] of Object.entries(clientReq.headers)) {
     const lowerKey = key.toLowerCase();
-    if (!hopByHopHeaders.includes(lowerKey)) {
+    if (!hopByHopHeaders.has(lowerKey)) {
       proxyHeaders[lowerKey] = value;
     }
   }
@@ -51,7 +52,7 @@ export const PassthroughServer = http.createServer((clientReq, clientRes) => {
   }, (proxyRes) => {
     const clientResHeaders = {};
     for (const [key, value] of Object.entries(proxyRes.headers)) {
-      if (!hopByHopHeaders.includes(key.toLowerCase())) {
+      if (!hopByHopHeaders.has(key.toLowerCase())) {
         clientResHeaders[key] = value;
       }
     }
@@ -75,3 +76,51 @@ export const PassthroughServer = http.createServer((clientReq, clientRes) => {
 
   clientReq.pipe(proxyReq);
 });
+
+type Result<T> = [err: Error, result: null] | [err: null, result: T];
+
+/**
+ * There's an issue with proxy-chain's handling of chunked requests when doing a direct passthrough.
+ * This workaround forwards the requests manually and returns the response
+ */
+export const makePassthrough = function ({ request, hostname, port }: PrepareRequestFunctionOpts): NonNullable<PrepareRequestFunctionResult['customResponseFunction']> {
+  return async () => {
+    const [err, proxyRes]: Result<http.IncomingMessage> = await new Promise((resolve) => {
+      const forward = http.request(
+        {
+          hostname,
+          port,
+          method: request.method,
+          path: request.url,
+          headers: request.headers,
+        },
+        (res) => resolve([null, res]),
+      );
+
+      forward.on("error", (err) => resolve([err, null]));
+      request.pipe(forward);
+    });
+
+    if (err) {
+      console.error(`Request failed "${err.name}": ${err.message}`)
+      throw err;
+    }
+    
+    const chunks: Buffer[] = [];
+    for await (const chunk of proxyRes) chunks.push(chunk);
+    const body = Buffer.concat(chunks);
+
+    const headers: IncomingHttpHeaders = {};
+    for (const [k, v] of Object.entries(proxyRes.headers)) {
+      if (!hopByHopHeaders.has(k.toLowerCase()) && v !== undefined) {
+        headers[k] = Array.isArray(v) ? v.join(",") : v;
+      }
+    }
+    
+    return {
+      statusCode: proxyRes.statusCode ?? 500,
+      headers: proxyRes.headers as Record<string, string>,
+      body,
+    };
+  };
+}
