+++ api/src/utils/proxy.ts
@@ -1,67 +1,7 @@
 import { env } from "../env.js";
 import { SessionService } from "../services/session.service.js";
-import { PrepareRequestFunctionOpts, PrepareRequestFunctionResult, Server } from "proxy-chain";
-import { PassthroughServer } from "./passthough-proxy.js";
-import http, { IncomingHttpHeaders } from "node:http";
-
-const hopByHop = new Set([
-  "connection",
-  "proxy-authenticate",
-  "proxy-authorization",
-  "keep-alive",
-  "te",
-  "trailers",
-  "transfer-encoding",
-  "upgrade",
-]);
-
-type Result<T> = [err: Error, result: null] | [err: null, result: T];
-
-/**
- * There's an issue with proxy-chain's handling of chunked requests when doing a direct passthrough.
- * This workaround forwards the requests manually and returns the response
- */
-const makePassthrough = function ({ request, hostname, port }: PrepareRequestFunctionOpts): NonNullable<PrepareRequestFunctionResult['customResponseFunction']> {
-  return async () => {
-    const [err, proxyRes]: Result<http.IncomingMessage> = await new Promise((resolve) => {
-      const forward = http.request(
-        {
-          hostname,
-          port,
-          method: request.method,
-          path: request.url,
-          headers: request.headers,
-        },
-        (res) => resolve([null, res]),
-      );
-
-      forward.on("error", (err) => resolve([err, null]));
-      request.pipe(forward);
-    });
-
-    if (err) {
-      console.error(`Request failed "${err.name}": ${err.message}`)
-      throw err;
-    }
-    
-    const chunks: Buffer[] = [];
-    for await (const chunk of proxyRes) chunks.push(chunk);
-    const body = Buffer.concat(chunks);
-
-    const headers: IncomingHttpHeaders = {};
-    for (const [k, v] of Object.entries(proxyRes.headers)) {
-      if (!hopByHop.has(k.toLowerCase()) && v !== undefined) {
-        headers[k] = Array.isArray(v) ? v.join(",") : v;
-      }
-    }
-    
-    return {
-      statusCode: proxyRes.statusCode ?? 500,
-      headers: proxyRes.headers as Record<string, string>,
-      body,
-    };
-  };
-}
+import { Server } from "proxy-chain";
+import { makePassthrough, PassthroughServer } from "./passthough-proxy.js";
 
 export class ProxyServer extends Server {
   public url: string;
